Start refactoring
