--- @jxnblk/gatsby-theme-mdx-blog/gatsby-browser.js
@@ -1,1 +1,0 @@
-export { wrapRootElement } from './src'
