--- @jxnblk/gatsby-theme-mdx-blog/gatsby-ssr.js
@@ -1,1 +1,0 @@
-export { wrapRootElement } from './src'
