--- @jxnblk/gatsby-theme-mdx-blog/src/container.js
@@ -1,10 +1,0 @@
-import React from 'react'
-import { css, Container } from 'theme-ui'
-
-export default props =>
-  <Container
-    {...props}
-    css={css({
-      variant: 'layout.container'
-    })}
-  />
