+++ @jxnblk/gatsby-theme-mdx-blog/src/gatsby-plugin-theme-ui/colors.js
