--- @jxnblk/gatsby-theme-mdx-blog/src/layout.js
@@ -1,6 +1,0 @@
-export default {
-  root: {},
-  header: {},
-  footer: {},
-  container: {}
-}
