+++ @jxnblk/gatsby-theme-mdx-blog/src/root.js
@@ -5,7 +5,6 @@ export const Root = props =>
   <Layout
     {...props}
     css={css({
-      variant: 'layout.root',
       fontFamily: 'body',
       fontWeight: 'body',
       lineHeight: 'body',
